Use a common base model to share common behaviour.
