+++ module/Geo/src/Model/Location.php
@@ -2,9 +2,9 @@
 
 namespace Rox\Geo\Model;
 
-use Illuminate\Database\Eloquent\Model;
+use Rox\Core\Model\AbstractModel;
 
-class Location extends Model
+class Location extends AbstractModel
 {
     /**
      * @var boolean
@@ -19,7 +19,7 @@ class Location extends Model
     /**
      * @var array
      */
-    protected $relationships = [
+    protected $ormRelationships = [
         'country',
     ];
 
@@ -28,27 +28,8 @@ class Location extends Model
         return $this->hasOne(Country::class, 'country', 'countryCode');
     }
 
-    public function getAttribute($key)
-    {
-        // The Eloquent implementation of getAttribute will first return the
-        // attribute of $key before checking if it has a relationship.
-        // We want the opposite of this because we want to define the 'country'
-        // key as a relationship to the geoname entity, even though the location
-        // table defines a 'country' column.
-        if (in_array($key, $this->relationships, true)) {
-            return $this->getRelationValue($key);
-        }
-
-        return parent::getAttribute($key);
-    }
-
     public function getCountryCodeAttribute()
     {
         return $this->attributes['country'];
-    }
-
-    public function __isset($key)
-    {
-        return parent::__isset($key) || in_array($key, $this->relationships, true);
     }
 }
